# shared/openapi/v1.json
{
  "components": {
    "schemas": {
      "BusinessCreateIn": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "contact_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Phone"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "display_name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          }
        },
        "required": [
          "display_name"
        ],
        "title": "BusinessCreateIn",
        "type": "object"
      },
      "BusinessMemberAddIn": {
        "description": "Body for `POST /businesses/me/members`.\n\nThe owner identifies a user by phone (the same identifier they use\nfor OTP login). `role` is constrained to manager/staff — owner is\nset during business creation and cannot be added or changed via\nthis endpoint.",
        "properties": {
          "role": {
            "enum": [
              "manager",
              "staff"
            ],
            "title": "Role",
            "type": "string"
          },
          "user_phone": {
            "title": "User Phone",
            "type": "string"
          }
        },
        "required": [
          "user_phone",
          "role"
        ],
        "title": "BusinessMemberAddIn",
        "type": "object"
      },
      "BusinessMemberDeleteOut": {
        "properties": {
          "ok": {
            "const": true,
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "BusinessMemberDeleteOut",
        "type": "object"
      },
      "BusinessMemberListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BusinessMemberOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "BusinessMemberListOut",
        "type": "object"
      },
      "BusinessMemberOut": {
        "properties": {
          "business_id": {
            "format": "uuid",
            "title": "Business Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/BusinessMemberRole"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "business_id",
          "user_id",
          "role",
          "created_at",
          "updated_at"
        ],
        "title": "BusinessMemberOut",
        "type": "object"
      },
      "BusinessMemberRole": {
        "description": "Access role for a user inside a business.\n\n- `owner`   — original creator; cannot be removed; can invite/remove others.\n- `manager` — can manage SKUs, members below them, and warehouse settings.\n- `staff`   — can record stock movements and use the chat surface.\n\nThe phase-1 1:1 invariant on `businesses.owner_id` survives — that\ncolumn is still the canonical default — but `business_members` is\nthe source of truth for access checks from session 10 onward.",
        "enum": [
          "owner",
          "manager",
          "staff"
        ],
        "title": "BusinessMemberRole",
        "type": "string"
      },
      "BusinessOut": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "contact_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Phone"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "owner_id": {
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "owner_id",
          "display_name",
          "description",
          "address",
          "contact_phone",
          "created_at",
          "updated_at"
        ],
        "title": "BusinessOut",
        "type": "object"
      },
      "BusinessUpdateIn": {
        "description": "Partial update. Only provided fields are applied.\n\n`contact_phone` accepts an explicit empty string or None to clear the\noverride, since a business might want to fall back to the owner's phone\nlater. We represent that as a sentinel in the service layer.",
        "properties": {
          "address": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "contact_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Phone"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          }
        },
        "title": "BusinessUpdateIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SteeringSide": {
        "description": "Steering-wheel / drive-side layout of a physical car.\n\nMongolia has a mixed fleet: Japanese imports are overwhelmingly RHD\n(steering on the right), while Korean / German / US / Chinese imports\nare LHD. Business coverage tables (session 5) use this as a filter so\na shop stocking LHD Tiguan parts doesn't get searches for RHD Prii.\n\nMapping from the XYP `wheelPosition` field:\n    \"Зүүн\"  (Mongolian for \"left\")  → LHD\n    \"Баруун\" (Mongolian for \"right\") → RHD",
        "enum": [
          "LHD",
          "RHD"
        ],
        "title": "SteeringSide",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VehicleBrandCoverageIn": {
        "description": "One entry in a business's coverage set.\n\n`year_start`/`year_end` are inclusive bounds on vehicle.build_year.\nEither can be NULL; NULL-NULL means \"every year this brand has ever\nbeen made\". `steering_side` NULL accepts both LHD and RHD.",
        "properties": {
          "steering_side": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SteeringSide"
              },
              {
                "type": "null"
              }
            ]
          },
          "vehicle_brand_id": {
            "format": "uuid",
            "title": "Vehicle Brand Id",
            "type": "string"
          },
          "year_end": {
            "anyOf": [
              {
                "maximum": 2100.0,
                "minimum": 1900.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year End"
          },
          "year_start": {
            "anyOf": [
              {
                "maximum": 2100.0,
                "minimum": 1900.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year Start"
          }
        },
        "required": [
          "vehicle_brand_id"
        ],
        "title": "VehicleBrandCoverageIn",
        "type": "object"
      },
      "VehicleBrandCoverageListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VehicleBrandCoverageOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "VehicleBrandCoverageListOut",
        "type": "object"
      },
      "VehicleBrandCoverageOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "steering_side": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SteeringSide"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "vehicle_brand_id": {
            "format": "uuid",
            "title": "Vehicle Brand Id",
            "type": "string"
          },
          "year_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year End"
          },
          "year_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year Start"
          }
        },
        "required": [
          "vehicle_brand_id",
          "year_start",
          "year_end",
          "steering_side",
          "created_at",
          "updated_at"
        ],
        "title": "VehicleBrandCoverageOut",
        "type": "object"
      },
      "VehicleBrandCoverageReplaceIn": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VehicleBrandCoverageIn"
            },
            "maxItems": 50,
            "title": "Items",
            "type": "array"
          }
        },
        "title": "VehicleBrandCoverageReplaceIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "iAuto API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/businesses": {
      "post": {
        "operationId": "create_business_v1_businesses_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a business profile for the authenticated user",
        "tags": [
          "businesses"
        ]
      }
    },
    "/v1/businesses/me": {
      "get": {
        "operationId": "get_my_business_v1_businesses_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Return the authenticated owner's business profile",
        "tags": [
          "businesses"
        ]
      },
      "patch": {
        "operationId": "update_my_business_v1_businesses_me_patch",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Partially update the authenticated owner's business profile",
        "tags": [
          "businesses"
        ]
      }
    },
    "/v1/businesses/me/members": {
      "get": {
        "operationId": "list_members_v1_businesses_me_members_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessMemberListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List members of the caller's business",
        "tags": [
          "businesses"
        ]
      },
      "post": {
        "operationId": "add_member_v1_businesses_me_members_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessMemberAddIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessMemberOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add a manager / staff member to the caller's business (owner only)",
        "tags": [
          "businesses"
        ]
      }
    },
    "/v1/businesses/me/members/{user_id}": {
      "delete": {
        "operationId": "remove_member_v1_businesses_me_members__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessMemberDeleteOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove a member from the caller's business (owner only)",
        "tags": [
          "businesses"
        ]
      }
    },
    "/v1/businesses/me/vehicle-brands": {
      "get": {
        "operationId": "list_vehicle_brands_v1_businesses_me_vehicle_brands_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleBrandCoverageListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List the business's current vehicle brand coverage",
        "tags": [
          "businesses"
        ]
      },
      "put": {
        "operationId": "replace_vehicle_brands_v1_businesses_me_vehicle_brands_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleBrandCoverageReplaceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleBrandCoverageListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace the business's vehicle brand coverage set",
        "tags": [
          "businesses"
        ]
      }
    },
...
  }
}